Fix computation for buffer route domain
The previous domain for pull rule was comparing the parent location of
the rules' source locations. It may work by chance in some cases, but
doesn't seem accurate.

We are looking for the rule that goes to the buffer, so we have to find
the routes of the rules which destination matches the buffer's location
or one of its parent location.

In fact, it matches the domain of "StockRule._get_rule()". If it
doesn't, the rule would not be applied.

As the rule destination is mandatory, we don't need anymore an explicit
check for the "manufacture" rules.