Add ddmrp_cron_actions_as_job
It makes calls to "cron_actions" run in queue jobs.

The jobs have an identity key with "identity_exact", meaning that only
one occurence of a job for the same buffer with the same arguments
(only_nfp) will be created at a time (e.g. when the state of a
stock.move is changed several times in the same transaction or in
a different transaction in a short timeframe).

It needs https://github.com/OCA/queue/pull/274 and
https://github.com/OCA/queue/pull/275