[IMP] Add imposed demand lead time to buffers.
When defined, the yellow zone will be computed as the ADU x Imposed Demand Lead Time, if the Imposed Demand Lead
Time is larger than the Replenishment Lead Time. This is particularily useful in situations of infrequent but
periodic demand. E.g. We receive a large order every 30 days, whereas the supplier takes 10 days to supply.

In this case the yellow zone must cover for the entire cycle of 30 days of demand.

In situations with infrequent demand the ADU tends to be very small, and every new order would be treated
as a spike, when in reality this is not an exceptional situation.