Invalidate computed non-stored fields in cron_actions
Fix tests:

* test_22_procure_recommended
* test_23_buffer_zones_all

It is possible that the invalidation is only required because of the
tests and would not be required at runtime. But this has been proposed
in the past: https://github.com/OCA/ddmrp/pull/59

However, this commit tries to invalidate the less possible fields to
limit the computations and data to fetch again.