[13.0] Move On-Hand status out of purchase order line
In the purchase order form view, the on hand status is visible in the
ordered lines tree view.
This information slows down the load of every large purchase order when
this information is not always usefull.

To improve on this. This change removes the on hand status from the
purchase order form.
And adds a smart button on the from to visualize this information.