[13.0][ADD] basic stock buffer tree view
The default stock.buffer tree view can become slow with a lot of
records.
The idea with this PR is to have a more basic stock.buffer tree view
with less fields that would be a little more responsive.

The existing tree view will be called Stock Buffer Details to
differentiate the two.

Unfortunately it does not seem to be much of a speed improvement, mainly
because of the `_compute_product_available_qty` function which is run
for fields that are too important to remove even in it's basic form.