Add ddmrp_packaging_product_replace
There seems to be a different outcome when doing a product replacement
with packaging set on the related stock buffers between v13 and v14. In
v13 the following test passes.
But on v14 the new stock buffer created has the packaging of the old
product set on them and the constraint `_check_product_packaging` raises
an error.