Fix planning chart display in ddmrp_history
When displaying some buffers there is an error coming up from the
bokeh library:

    Error: invalid bbox {left: NaN, top: NaN, right: NaN, bottom: NaN}

And the planning chart would not be displayed.

Looking at the data used for generating the chart, the green, yellow and red are
populated with zeros. This lead to an y axis of 0 to 0.

But also the two positions data (net flow and on hand) were sometimes populated
with negative values.

This fixed the problem but maybe those values generated by the buffer were
already incorrect.