Fix crash on execution chart display
The ddmrp view crashes with this error

    Error: invalid bbox {left: NaN, top: NaN, right: NaN, bottom: NaN}

This one is on the Execution Chart and it is caused because the y axis
of the graph has the same value for the bottom and top value.

Which I guess for the Bokeh library it result to a non existing chart.
To not have a top value equal to zero fixes the crash.