[FIX] deprecated lint disable estatement
`pylint:disable` throws a deprecated warning, however the new oca-disable
does not work as expected. The workaround is to disable the check for
oca-checks-odoo.

The same check is done by pytlint_odoo and in that case the old `pylint:disable`
syntax do have an effect.

See https://github.com/OCA/oca-addons-repo-template/issues/233 for details.