[FIX] ddmrp: Remove mrp_move from qualified demand calculation
Qualified demand should only consider actual demand, such as outgoing stock moves or any entity that refers to something confirmed and real.
MRP Move is an entity similar to Stock Demand Estimate where we forecast future demand and should not be considered as qualified demand.

Stock buffers should consider forecasted future demand via the ADU calculation, sizing the buffer according to that demand. But, then, only at the moment that this demand is confirmed it should be considered as qualified demand.